[Feat] Code Refactoring

+++ src/lib/components/pages/post/postController.tsx
@@ -1,17 +1,18 @@
 "use client";
 
-import { Content } from "@/lib/utils/constants/postProps";
+import Content from "@/lib/types/content";
 import IconButton from "@/lib/components/iconButton";
 import DeleteIcon from "@/lib/icons/delete";
 import EditIcon from "@/lib/icons/edit";
 import UpArrowIcon from "@/lib/icons/upArrow";
 import { PostAction, PostActionType } from "@/lib/utils/reducers/postReducer";
-import "@/style/post.css";
 import { Dispatch } from "@reduxjs/toolkit";
 import { useSession } from "next-auth/react";
 import { useRouter } from "next/navigation";
 import { useDispatch } from "react-redux";
-import { useEffect, useState } from "react";
+import { useEffect } from "react";
+import useFetchData from "@/lib/hooks/useFetchData";
+import "@/style/post.css";
 
 const size = "2rem";
 
@@ -20,32 +21,24 @@ interface Props {
 }
 
 const PostController: React.FC<Props> = ({ post }) => {
-  const { data } = useSession();
-  const router = useRouter();
   const dispatch: Dispatch<PostAction> = useDispatch();
+  const { data: session } = useSession();
+  const router = useRouter();
 
-  const [role, setRole] = useState("");
+  const { data: role, fetchData } = useFetchData(
+    "/api/auth/users",
+    "guest",
+    false
+  );
 
   useEffect(() => {
-    const fetchData = async () => {
-      const response = await fetch("/api/auth/user", {
+    if (session?.user?.email) {
+      fetchData({
         method: "POST",
-        body: JSON.stringify({ email: data?.user?.email }),
+        body: JSON.stringify({ email: session?.user?.email }),
       });
-
-      const jsonData = await response.json();
-
-      switch (response.status) {
-        case 200:
-          setRole(jsonData["role"]);
-          break;
-      }
-    };
-
-    if (data?.user?.email) {
-      fetchData();
     }
-  }, [data?.user?.email]);
+  }, [session?.user?.email, fetchData]);
 
   return (
     <div className="post-controller">
@@ -68,56 +61,53 @@ const PostController: React.FC<Props> = ({ post }) => {
             onClick={() => {
               dispatch({
                 type: PostActionType.SET_POST,
-                payload: {
-                  id: post?.id,
-                  title: post?.title,
-                  description: post?.description,
-                  thumbnail: post?.thumbnail,
-                  content: post?.content,
-                  topic: post?.topic,
-                  series: post?.series,
-                  published: post?.published,
-                  author: post?.author,
-                  _count: post?._count,
-                },
+                payload: post,
               });
-              router.push("/write");
+              router.push("/admin/write");
             }}
           >
             <EditIcon width={size} height={size} />
           </IconButton>
-          <IconButton
-            description="삭제하기"
-            onClick={() => {
-              const flag = confirm("정말로 삭제하시겠습니까?");
-
-              if (!flag) return;
-
-              const fetchData = async () => {
-                const response = await fetch("/api/post", {
-                  method: "delete",
-                  body: JSON.stringify({ post }),
-                });
-
-                switch (response.status) {
-                  case 200:
-                    alert("삭제되었습니다");
-                    window.location.href = "/";
-                    break;
-                  default:
-                    alert("삭제 안됨");
-                    break;
-                }
-              };
-
-              fetchData();
-            }}
-          >
-            <DeleteIcon width={size} height={size} />
-          </IconButton>
+          <DeleteButton post={post} />
         </>
       ) : null}
     </div>
+  );
+};
+
+//----------------------------------------------------------
+
+const DeleteButton: React.FC<Props> = ({ post }) => {
+  const {
+    data: success,
+    isLoading,
+    fetchData,
+  } = useFetchData("/api/admin/posts", false, false);
+
+  useEffect(() => {
+    if (success) {
+      window.location.href = "/";
+    }
+  }, [success]);
+
+  return (
+    <IconButton
+      description="삭제하기"
+      onClick={() => {
+        const flag = confirm("정말로 삭제하시겠습니까?");
+
+        if (!flag) return;
+
+        if (!isLoading) {
+          fetchData({
+            method: "delete",
+            body: JSON.stringify({ posts: post }),
+          });
+        }
+      }}
+    >
+      <DeleteIcon width={size} height={size} />
+    </IconButton>
   );
 };
